pyproject.toml
[project]
name = "pnftrading-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn",
    "httpx",
    "requests",
    "paho-mqtt",
]
